requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-multipart
python-jose
passlib[bcrypt]
sqlalchemy==1.4.*

pytest
pytest-xdist
pytest-cov
httpx
ruff
